add skip_untar option for start_diag script

+++ diag/python/regression/scripts/start_diag.sh
@@ -10,6 +10,10 @@ else
         slot=$2
     else
         arch=amd64
+        if [[ $1 == "skip_untar" ]]
+        then
+            skip_untar=1
+        fi
     fi
 fi
 
@@ -18,17 +22,21 @@ echo "-------------------"
 echo "Preparing diag environment"
 DIAG_DIR=/home/diag/diag
 
-echo "Untar ASIC lib"
 ASIC_IMG=/home/diag/nic.tar.gz
 asic_type=$(grep "ASIC_TYPE" $DIAG_DIR/python/regression/scripts/dft_profile_mtp | cut -d "=" -f 2)
 asic=$(echo $asic_type | awk '{print tolower($0)}')
 echo "ASIC: $asic"
-chmod -R 755 $DIAG_DIR/asic_all/$asic/
-tar xf $ASIC_IMG -C $DIAG_DIR/asic_all/$asic/
-cp -r $DIAG_DIR/asic_all/$asic/nic/* $DIAG_DIR/asic_all/$asic/
-cp $DIAG_DIR/asic_all/$asic/asic_src/ip/cosim/tclsh/.git_rev.tcl $DIAG_DIR/asic_all/$asic/asic_version.txt
-rm -rf $DIAG_DIR/asic_all/$asic/nic
-sync;sync;sync
+if [[ -n $skip_untar ]]
+then
+    echo "Using existing ASIC lib"
+else
+    echo "Untar ASIC lib"
+    chmod -R 755 $DIAG_DIR/asic_all/$asic/
+    tar xf $ASIC_IMG -C $DIAG_DIR/asic_all/$asic/
+    cp -r $DIAG_DIR/asic_all/$asic/nic/* $DIAG_DIR/asic_all/$asic/
+    cp $DIAG_DIR/asic_all/$asic/asic_src/ip/cosim/tclsh/.git_rev.tcl $DIAG_DIR/asic_all/$asic/asic_version.txt
+    rm -rf $DIAG_DIR/asic_all/$asic/nic
+fi
 
 mkdir -p $DIAG_DIR/log/
